Add interfaces and implementations for action properties
For the action server/client it is needed that Generics are present for
the action messages. Each of the three action classes now implements an
interface which provides properties to action related fields, like
Header, GoalId and GoalStatus. The inner message (goal, result, feedback)
is also accessible through that interface.